refactor(workout-builder): move useWorkoutSession hook to workout-session feature for better modularity and organization feat(workout-session): implement useWorkoutSession hook and workout session store to manage workout state and session data effectively
