+++ src/features/workout-builder/index.ts
@@ -1,5 +1,5 @@
 export { WorkoutStepper } from "./ui/workout-stepper";
 export { useWorkoutStepper } from "./model/use-workout-stepper";
-export { useWorkoutSession } from "./model/use-workout-session";
+export { useWorkoutSession } from "../workout-session/model/use-workout-session";
 export { EQUIPMENT_CONFIG, getEquipmentByValue, getEquipmentLabel } from "./model/equipment-config";
 export type { WorkoutBuilderState, WorkoutBuilderStep, EquipmentItem } from "./types";
