+++ src/features/workout-session/model/use-workout-session.ts
@@ -1,0 +1,7 @@
+"use client";
+
+import { useWorkoutSessionStore } from "./workout-session.store";
+
+export function useWorkoutSession() {
+  return useWorkoutSessionStore();
+}
